# backend/src/main/java/com/batteryrecycling/traceability/attachment/AttachmentDtos.java
package com.batteryrecycling.traceability.attachment;

import java.time.LocalDateTime;

public final class AttachmentDtos {
    private AttachmentDtos() {
    }

    public record AttachmentDto(
            Long id,
            String fileName,
            String fileExt,
            Long fileSizeBytes,
            String bindingStatus,
            LocalDateTime expiresAt
    ) {
    }
}