# contracts/api/openapi-first-slice.yaml
openapi: 3.0.3
info:
  title: Battery Recycling Traceability First Slice API
  version: 0.1.0
  description: 第一切片 API 设计契约，用于技术设计评审，不等同于已冻结生产接口。
  license:
    name: Course Project Internal Use
    url: https://example.com/course-project-internal-use-license
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /acceptances/pending:
    get:
      tags: [Acceptance]
      summary: 查询待验收任务
      operationId: listPendingAcceptances
      x-traceability:
        fr: [FR-C4-011, FR-C4-012, FR-C4-014]
        ac: [AC-005, AC-006, AC-008]
        tc: [TC-C4-001, TC-C4-007, TC-C4-008]
        ui: [UI-C4-007]
        permission: acceptance:create
        module: MOD-ACCEPTANCE
        db: [battery, recycle_batch_battery]
      responses:
        '200':
          $ref: '#/components/responses/BatteryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /batteries/{id}/acceptances:
    post:
      tags: [Acceptance]
      summary: 保存验收记录
      operationId: createAcceptance
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      x-traceability:
        fr: [FR-C4-011, FR-C4-012, FR-C4-014, FR-C4-019]
        ac: [AC-005, AC-006, AC-008, AC-012]
        tc: [TC-C4-001, TC-C4-007, TC-C4-008, TC-C4-016]
        ui: [UI-C4-008]
        permission: acceptance:create
        module: MOD-ACCEPTANCE
        db: [acceptance_record, battery, recycle_batch, lifecycle_event, audit_log]
        transaction: saveAcceptance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptanceCreateRequest'
      responses:
        '200':
          $ref: '#/components/responses/AcceptanceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /batteries/{id}/acceptance-supplements:
    post:
      tags: [Acceptance]
      summary: 补充验收资料
      operationId: supplementAcceptance
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      x-traceability:
        fr: [FR-C4-013, FR-C4-019]
        ac: [AC-007, AC-012]
        tc: [TC-C4-007, TC-C4-014]
        ui: [UI-C4-009]
        permission: acceptance:supplement
        module: MOD-ACCEPTANCE
        db: [acceptance_supplement, battery, lifecycle_event, audit_log]
        transaction: supplementAcceptance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptanceSupplementRequest'
      responses:
        '200':
          $ref: '#/components/responses/BatteryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /acceptance-records/{id}:
    delete:
      tags: [Acceptance]
      summary: 删除保护入口
      operationId: deleteAcceptanceRecord
      parameters:
        - $ref: '#/components/parameters/PathId'
      x-traceability:
        fr: [FR-C4-022]
        ac: [AC-014]
        tc: [TC-C4-013]
        ui: [UI-C4-014]
        permission: authenticated
        module: MOD-AUDIT
        db: [acceptance_record, audit_log]
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /attachments:
    post:
      tags: [System]
      summary: 上传业务附件
      operationId: uploadAttachment
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      x-traceability:
        fr: [FR-C4-003, FR-C4-013]
        ac: [AC-001, AC-007]
        tc: [TC-C4-007, TC-C4-015]
        ui: [UI-C4-004, UI-C4-009]
        permission: attachment:upload
        module: MOD-ATTACHMENT
        db: [business_attachment, idempotency_record]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AttachmentUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/AttachmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /attachments/{id}/download:
    get:
      tags: [System]
      summary: 下载业务附件
      operationId: downloadAttachment
      parameters:
        - $ref: '#/components/parameters/PathId'
      x-traceability:
        fr: [FR-C4-003, FR-C4-013]
        ac: [AC-001, AC-007]
        tc: [TC-C4-007, TC-C4-015]
        ui: [UI-C4-004, UI-C4-009]
        permission: attachment:read
        module: MOD-ATTACHMENT
        db: [business_attachment]
      responses:
        '200':
          description: 附件文件流
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    BadRequest:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 认证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 状态冲突或重复提交
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    BatteryResponse:
      description: 电池响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BatteryEnvelope'
...
    BatteryListResponse:
      description: 电池列表响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BatteryListEnvelope'
    AcceptanceResponse:
      description: 验收响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AcceptanceEnvelope'
...
    AttachmentResponse:
      description: 附件响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AttachmentEnvelope'
  schemas:
    Envelope:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          required: [timestamp]
          properties:
            timestamp:
              type: string
              format: date-time
            data:
              type: object
              nullable: true
...
    BatteryEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Battery'
...
    BatteryListEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Battery'
    Battery:
      type: object
      required: [id, systemTraceCode, batteryType, batteryChemistry, lifecycleStatus, duplicateStatus]
      properties:
        id:
          type: integer
          format: int64
        enterpriseId:
          type: integer
          format: int64
        systemTraceCode:
          type: string
        originalCode:
          type: string
          maxLength: 100
          nullable: true
        batteryType:
          type: string
          enum: [PACK]
        batteryModel:
          type: string
          nullable: true
        manufacturer:
          type: string
          nullable: true
        batteryChemistry:
          type: string
        nominalCapacity:
          type: number
          nullable: true
        productionDate:
          type: string
          format: date
          nullable: true
        currentResponsibleEnterpriseId:
          type: integer
          format: int64
        lifecycleStatus:
          type: string
        duplicateStatus:
          type: string
        version:
          type: integer
...
    AcceptanceCreateRequest:
      type: object
      required: [acceptanceResult, identityCheckResult, appearanceCheckResult, documentCheckResult]
      properties:
        acceptanceResult:
          type: string
          enum: [PASS, NEED_SUPPLEMENT, REJECT]
        identityCheckResult:
          type: string
          maxLength: 40
        appearanceCheckResult:
          type: string
          maxLength: 40
        documentCheckResult:
          type: string
          maxLength: 40
        acceptanceNote:
          type: string
          maxLength: 500
    AcceptanceEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                acceptanceRecordId:
                  type: integer
                  format: int64
                batteryStatus:
                  type: string
    AcceptanceSupplementRequest:
      type: object
      properties:
        supplementNote:
          type: string
          maxLength: 500
        attachmentIds:
          type: array
          items:
            type: integer
            format: int64
...
    AttachmentUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    AttachmentEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Attachment'
    Attachment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fileName:
          type: string
        fileExt:
          type: string
        fileSizeBytes:
          type: integer
          format: int64
        bindingStatus:
          type: string
          enum: [TEMP, BOUND]
        expiresAt:
          type: string
          format: date-time
...